Time: 252 ms (64.67%), Space: 65.7 MB (65.48%) - LeetHub

+++ 1014-k-closest-points-to-origin/1014-k-closest-points-to-origin.cpp
@@ -2,14 +2,14 @@ class Solution {
 public:
     vector<vector<int>> kClosest(vector<vector<int>>& points, int k) {
         vector<vector<int>> ans;
-        priority_queue<pair<int,pair <int,int>>> maxHeap;
+        priority_queue<pair<int,pair <long long,long long>>> maxHeap;
         
         for (int i=0; i<points.size(); i++){
            
-            int x_square = points[i][0]*points[i][0];
-            int y_square = points[i][1]*points[i][1];
+            long long x_square = points[i][0]*points[i][0];
+            long long y_square = points[i][1]*points[i][1];
 
-            int value = x_square + y_square;
+            long long value = x_square + y_square;
 
             maxHeap.push({value,{points[i][0],points[i][1]}});
 
@@ -17,7 +17,7 @@ public:
         }
 
         while (!maxHeap.empty()){
-            ans.push_back({maxHeap.top().second.first,maxHeap.top().second.second});
+            ans.push_back({(int)maxHeap.top().second.first,(int)maxHeap.top().second.second});
             maxHeap.pop();
         }
